AnsibleModuleDev/requirements.txt
ansible-core
pypsrp
smbprotocol